+++ Sources/KeyboardKit/Layout/KeyboardLayout+Configuration.swift
@@ -11,23 +11,18 @@ import SwiftUI
 
 public extension KeyboardLayout {
     
-    /**
-     This struct defines keyboard layout configs for various
-     device types.
-     
-     The standard parameterless configuration can be changed
-     to affect the global defaults.
-     */
+    /// This type defines layout configs for various devices.
+    ///
+    /// The standard configurations can be changed to affect
+    /// the global defaults.
     struct Configuration: Equatable {
         
-        /**
-         Create a new layout configuration.
-         
-         - Parameters:
-           - buttonCornerRadius: The corner radius of a keyboard button in the keyboard.
-           - buttonInsets: The edge insets of a keyboard button in the keyboard.
-           - rowHeight: The total height incl. insets, of a row in the keyboard.
-         */
+        /// Create a new layout configuration.
+        ///
+        /// - Parameters:
+        ///   - buttonCornerRadius: The button corner radius.
+        ///   - buttonInsets: The button edge insets.
+        ///   - rowHeight: The total row height including insets.
         public init(
             buttonCornerRadius: CGFloat,
             buttonInsets: EdgeInsets,
@@ -71,8 +66,8 @@ public extension KeyboardLayout.Configuration {
         orientation: InterfaceOrientation
     ) -> Self {
         switch device {
-        case .pad: return standardPad(forScreenSize: size, orientation: orientation)
-        default: return standardPhone(forScreenSize: size, orientation: orientation)
+        case .pad: standardPad(forScreenSize: size, orientation: orientation)
+        default: standardPhone(forScreenSize: size, orientation: orientation)
         }
     }
 
@@ -130,83 +125,73 @@ public extension KeyboardLayout.Configuration {
     
     // MARK: - Standard Configuration Properties
 
-    /**
-     The standard config for an iPad in portait.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for an iPad in portait.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPad = Self(
         buttonCornerRadius: 5,
         buttonInsets: .init(horizontal: 6, vertical: 4),
-        rowHeight: standardPadRowHeight)
+        rowHeight: standardPadRowHeight
+    )
 
-    /**
-     The standard config for an iPad in landscape.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for an iPad in landscape.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPadLandscape = Self(
         buttonCornerRadius: 7,
         buttonInsets: .init(horizontal: 7, vertical: 6),
-        rowHeight: standardPadLandscapeRowHeight)
+        rowHeight: standardPadLandscapeRowHeight
+    )
 
-    /**
-     The standard config for a large iPad Pro in portrait.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for large iPad Pros in portrait.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPadProLarge = Self(
         buttonCornerRadius: 6,
         buttonInsets: .init(horizontal: 4, vertical: 4),
-        rowHeight: standardPadProLargeRowHeight)
+        rowHeight: standardPadProLargeRowHeight
+    )
 
-    /**
-     The standard config for a large iPad Pro in landscape.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for large iPad Pros in landscape.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPadProLargeLandscape = Self(
         buttonCornerRadius: 8,
         buttonInsets: .init(horizontal: 7, vertical: 5),
-        rowHeight: standardPadProLargeLandscapeRowHeight)
+        rowHeight: standardPadProLargeLandscapeRowHeight
+    )
 
-    /**
-     The standard config for an iPhone in portrait.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for an iPhone in portrait.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPhone = Self(
         buttonCornerRadius: 5,
         buttonInsets: .init(horizontal: 3, vertical: 6),
-        rowHeight: standardPhoneRowHeight)
+        rowHeight: standardPhoneRowHeight
+    )
 
-    /**
-     The standard config for an iPhone in landscape.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for an iPhone in landscape.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPhoneLandscape = Self(
         buttonCornerRadius: 5,
         buttonInsets: .init(horizontal: 3, vertical: 4),
-        rowHeight: standardPhoneLandscapeRowHeight)
+        rowHeight: standardPhoneLandscapeRowHeight
+    )
 
-    /**
-     The standard config for an iPhone Pro Max in portrait.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for iPhone Pro Max in portrait.
     static var standardPhoneProMax = Self(
         buttonCornerRadius: 5,
         buttonInsets: .init(horizontal: 3, vertical: 5.5),
-        rowHeight: standardPhoneProMaxRowHeight)
+        rowHeight: standardPhoneProMaxRowHeight
+    )
 
-    /**
-     The standard config for an iPhone Pro Max in landscape.
-
-     You can change this value to affect the global default.
-     */
+    /// The standard config for iPhone Pro Max in landscape.
+    ///
+    /// You can set this config to affect the global default.
     static var standardPhoneProMaxLandscape = Self(
         buttonCornerRadius: 5,
         buttonInsets: .init(horizontal: 3, vertical: 4),
-        rowHeight: standardPhoneProMaxLandscapeRowHeight)
+        rowHeight: standardPhoneProMaxLandscapeRowHeight
+    )
 }