+++ Sources/KeyboardKit/Layout/View+KeyboardLayoutItemSize.swift
@@ -10,17 +10,15 @@ import SwiftUI
 
 public extension View {
     
-    /**
-     Apply a keyboard layout item size to the view.
-     
-     The `edgeInsets` parameter can be used to add intrinsic
-     edge insets within the interactable button area.
-     
-     - Parameters:
-       - item: The layout item to apply.
-       - rowWidth: The total row width (often screen width).
-       - inputWidth: The width in points of an input item.
-     */
+    /// Apply a keyboard layout item size to the view.
+    ///
+    /// The `edgeInsets` can be used to add intrinsic insets
+    /// within the interactable button area.
+    ///
+    /// - Parameters:
+    ///   - item: The layout item to apply.
+    ///   - rowWidth: The total row width (often screen width).
+    ///   - inputWidth: The width in points of an input item.
     func keyboardLayoutItemSize(
         for item: KeyboardLayout.Item,
         rowWidth: CGFloat,